Refactor : AuthService인터페이스 생성
